vibrato for soundfonts + add getVibratoOscillator helper
