+++ packages/superdough/synth.mjs
@@ -1,6 +1,6 @@
 import { midiToFreq, noteToMidi } from './util.mjs';
 import { registerSound, getAudioContext } from './superdough.mjs';
-import { gainNode, getADSRValues, getParamADSR, getPitchEnvelope } from './helpers.mjs';
+import { gainNode, getADSRValues, getParamADSR, getPitchEnvelope, getVibratoOscillator } from './helpers.mjs';
 import { getNoiseMix, getNoiseOscillator } from './noise.mjs';
 
 const mod = (freq, range = 1, type = 'sine') => {
@@ -110,8 +110,6 @@ export function getOscillator(s, t, value) {
     n: partials,
     note,
     freq,
-    vib = 0,
-    vibmod = 0.5,
     noise = 0,
     // fm
     fmh: fmHarmonicity = 1,
@@ -181,17 +179,7 @@ export function getOscillator(s, t, value) {
   }
 
   // Additional oscillator for vibrato effect
-  let vibratoOscillator;
-  if (vib > 0) {
-    vibratoOscillator = getAudioContext().createOscillator();
-    vibratoOscillator.frequency.value = vib;
-    const gain = getAudioContext().createGain();
-    // Vibmod is the amount of vibrato, in semitones
-    gain.gain.value = vibmod * 100;
-    vibratoOscillator.connect(gain);
-    gain.connect(o.detune);
-    vibratoOscillator.start(t);
-  }
+  let vibratoOscillator = getVibratoOscillator(o.detune, value, t);
 
   // pitch envelope
   if (value.penv) {
